create algo/internal/utils/files for loadFilenames function
