+++ packages/algob/src/lib/files.ts
@@ -1,11 +1,9 @@
 import { getPathFromDirRecursive } from "@algo-builder/runtime";
 import { BuilderError, ERRORS, types } from "@algo-builder/web";
 import fs from "fs-extra";
-import glob from "glob";
 import path from "path";
 
 import { ASSETS_DIR } from "../internal/core/project-structure";
-import { cmpStr } from "./comparators";
 
 function normalizePaths (mainPath: string, paths: string[]): string[] {
   return paths.map(n => path.relative(mainPath, n));
@@ -49,28 +47,4 @@ export function loadEncodedTxFromFile (fileName: string): Uint8Array | undefined
     if (types.isFileError(e) && e?.errno === -2) { return undefined; } // handling a not existing file
     throw e;
   }
-}
-
-/**
- * Load .js, .ts files from /scripts (default) directory
- * @param directory directory to load files from
- * @param taskType task type (eg. test)
- * @returns array of paths as string eg. ['scripts/file1.js', 'scripts/file2.js', ..]
- */
-export function loadFilenames (directory: string, taskType?: string): string[] {
-  if (!fs.existsSync(directory)) {
-    if (taskType === "test") {
-      throw new BuilderError(ERRORS.BUILTIN_TASKS.TESTS_DIRECTORY_NOT_FOUND, {
-        directory
-      });
-    } else {
-      throw new BuilderError(ERRORS.BUILTIN_TASKS.SCRIPTS_DIRECTORY_NOT_FOUND, {
-        directory
-      });
-    }
-  }
-
-  return glob.sync(path.join(directory, "*.js"))
-    .concat(glob.sync(path.join(directory, "*.ts")))
-    .sort(cmpStr);
 }
