+++ packages/algob/test/builtin-tasks/deploy.ts
@@ -4,7 +4,7 @@ import * as fs from "fs";
 
 import { executeDeployTask as executeDeployTaskNoCLI } from "../../src/builtin-tasks/deploy";
 import { TASK_DEPLOY, TASK_RUN } from "../../src/builtin-tasks/task-names";
-import { loadFilenames } from '../../src/lib/files';
+import { loadFilenames } from '../../src/internal/util/files';
 import { loadCheckpoint } from "../../src/lib/script-checkpoints";
 import { expectBuilderErrorAsync } from "../helpers/errors";
 import { testFixtureOutputFile, useCleanFixtureProject } from "../helpers/project";
